restructure & remove unused
